bfsync commit: avoid problems with identical inodes during commit
In some (very rare) cases, bfsyncfs will copy-on-write the file associated with
the inode without modifying the file data or inode attributes.

The "fix unchanged inodes" pass is designed to reset the inode to the old
version in this case, before committing, which avoids problems with the actual
diff created later on.

Signed-off-by: Stefan Westerfeld <stefan@space.twc.de>